refactor: Move guest_config test data into module
Minimizes duplicating the creation of test data,
specifically for CustomCpuTemplate and CpuConfiguration.
